remove some files, fix the lint_test.go
