+++ pkg/lint/ignore/ignorer.go
@@ -3,6 +3,8 @@ package ignore
 import (
 	"helm.sh/helm/v3/pkg/lint/support"
 	"log/slog"
+	"path/filepath"
+	"strings"
 )
 
 type Ignorer struct {
@@ -10,6 +12,11 @@ type Ignorer struct {
 	Rules      []Rule
 	logger     *slog.Logger
 	RuleLoader *RuleLoader
+}
+
+type PathlessRule struct {
+	RuleText    string
+	MessageText string
 }
 
 // Ignorer is used to create the ignorer object that contains
@@ -48,4 +55,35 @@ func (i *Ignorer) ShouldKeepError(err error) bool {
 
 	// if we can't find a reason to discard it, we keep it
 	return true
+}
+
+type MatchesErrors interface {
+	Match(string) *RuleMatch
+}
+
+type RuleMatch struct {
+	ErrText  string
+	RuleText string
+}
+
+func (rm RuleMatch) LogAttrs() slog.Attr {
+	return slog.Group("rule_match", slog.String("err_text", rm.ErrText), slog.String("rule_text", rm.RuleText))
+}
+
+// Match errors that have no file path in their body with ignorer rules.
+// An examples of errors with no file path in their body is chart metadata errors `chart metadata is missing these dependencies`
+func (pr PathlessRule) Match(errText string) *RuleMatch {
+	ignorableError := pr.MessageText
+	parts := strings.SplitN(ignorableError, ":", 2)
+	prefix := strings.TrimSpace(parts[0])
+
+	if match, _ := filepath.Match(ignorableError, errText); match {
+		return &RuleMatch{ErrText: errText, RuleText: pr.RuleText}
+	}
+
+	if matched, _ := filepath.Match(prefix, errText); matched {
+		return &RuleMatch{ErrText: errText, RuleText: pr.RuleText}
+	}
+
+	return nil
 }
